+++ api/src/main/java/io/strimzi/plugin/security/profiles/impl/RestrictedPodSecurityProvider.java
@@ -96,6 +96,11 @@ public class RestrictedPodSecurityProvider extends BaselinePodSecurityProvider {
     }
 
     @Override
+    public SecurityContext kafkaBridgeInitContainerSecurityContext(ContainerSecurityProviderContext context) {
+        return createRestrictedContainerSecurityContext(context);
+    }
+
+    @Override
     public SecurityContext kafkaConnectBuildContainerSecurityContext(ContainerSecurityProviderContext context) {
         if (context != null
                 && context.userSuppliedSecurityContext() != null)    {